requirements.txt
pandas>=2.1
numpy>=1.26
pyyaml>=6.0
requests>=2.31
tqdm>=4.66
scikit-learn>=1.3
yfinance>=0.2
transformers>=4.40
torch>=2.1
sentencepiece>=0.1.99
langdetect>=1.0.9
beautifulsoup4>=4.12
python-dateutil>=2.8
joblib>=1.3
python-telegram-bot>=20.7
openai>=1.40.0
feedparser>=6.0
